Added warning to  "AddToPlaylist()" in users -> Preferable to use IPlaylist reference
Added basic work for playlist permission checks on delete. Requires a
rework of the internal permissions checking

+++ E.Deezer/Api/Playlist.cs
@@ -181,7 +181,7 @@ namespace E.Deezer.Api
                 RequestParameter.GetNewQueryStringParameter("songs", aTrackIds)
             };
 
-            return Client.Delete("playlist/{playlist_id}/tracks", parms, DeezerPermissions.ManageLibrary);
+            return Client.Delete("playlist/{playlist_id}/tracks", parms, DeezerPermissions.ManageLibrary | DeezerPermissions.DeleteLibrary);
         }
